// src/generated/api/models/PostCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostCreateDto
 */
export interface PostCreateDto {
    /**
     * 
     * @type {string}
     * @memberof PostCreateDto
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof PostCreateDto
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof PostCreateDto
     */
    type?: PostCreateDtoTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof PostCreateDto
     */
    imageUrl?: string;
}


/**
 * @export
 */
export const PostCreateDtoTypeEnum = {
    Public: 'PUBLIC',
    Private: 'PRIVATE',
    Draft: 'DRAFT'
} as const;
export type PostCreateDtoTypeEnum = typeof PostCreateDtoTypeEnum[keyof typeof PostCreateDtoTypeEnum];


/**
 * Check if a given object implements the PostCreateDto interface.
 */
export function instanceOfPostCreateDto(value: object): value is PostCreateDto {
    if (!('title' in value) || value['title'] === undefined) return false;
    return true;
}

export function PostCreateDtoFromJSON(json: any): PostCreateDto {
    return PostCreateDtoFromJSONTyped(json, false);
}

export function PostCreateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostCreateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'content': json['content'] == null ? undefined : json['content'],
        'type': json['type'] == null ? undefined : json['type'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
    };
}

export function PostCreateDtoToJSON(json: any): PostCreateDto {
    return PostCreateDtoToJSONTyped(json, false);
}

export function PostCreateDtoToJSONTyped(value?: PostCreateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'content': value['content'],
        'type': value['type'],
        'imageUrl': value['imageUrl'],
    };
}